# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestión de Pistas",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/"
        }
    ],
    "paths": {
...
        "/api/get-user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Obtener usuario y ver información de este",
                "operationId": "8676e42cf7da7995603730250593a5db",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "\n    *           $user (Object)"
                    },
                    "400": {
                        "description": "\n    *           Este usuario no existe"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Obtener usuario y ver información de este",
                "operationId": "6061bd3baaa2141c0499ae28e3d36426",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "\n    *           $user (Object)"
                    },
                    "400": {
                        "description": "\n    *           Este usuario no existe"
                    },
                    "403": {
                        "description": "\n    *           Esta operación solo lo puede hacer un administrador o un moderador"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/upload-image": {
            "post": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Actualizar la foto de perfil del usuario",
                "operationId": "51509ba5cd5870c079c2d285b93196dd",
                "parameters": [
                    {
                        "name": "foto_perfil",
                        "in": "query",
                        "description": "Foto de perfil del usuario",
                        "required": true,
                        "schema": {
                            "type": "image"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "\n    *           Foto actulizada con éxito\n    *           $path (string)"
                    },
                    "422": {
                        "description": "\n    *           Foto no válida"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/upload-image/{id}": {
            "post": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Actualizar la foto de perfil del usuario por su id",
                "operationId": "885ef9b2eee27cb3ece5613be297c105",
                "parameters": [
                    {
                        "name": "foto_perfil",
                        "in": "query",
                        "description": "Foto de perfil del usuario",
                        "required": true,
                        "schema": {
                            "type": "image"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "\n    *           Foto actulizada con éxito\n    *           $path (string)"
                    },
                    "400": {
                        "description": "\n    *           Usuario no existe"
                    },
                    "422": {
                        "description": "\n    *           Foto no válida"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/delete-image": {
            "delete": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Eliminar la foto de perfil del usuario",
                "operationId": "5f17340c49c3c21af21719422872b543",
                "responses": {
                    "202": {
                        "description": "Se ha eliminado la foto correctamente"
                    },
                    "400": {
                        "description": "No se ha podido eliminar la foto porque no existe"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/delete-image/{id}": {
            "delete": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Eliminar la foto de perfil del usuario",
                "operationId": "e4024e9ef2a607fb345369ff9a3fced7",
                "responses": {
                    "202": {
                        "description": "Se ha eliminado la foto correctamente"
                    },
                    "400": {
                        "description": "\n    *         Este usuario no existe\n    *         No se ha podido eliminar la foto porque no existe"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/get-image": {
            "get": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Obtener foto de perfil del usuario",
                "operationId": "ace29131851c4fc901df6f1de04472c4",
                "responses": {
                    "202": {
                        "description": "\n    *           $path_image (string)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/get-warnings": {
            "get": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Obtener las advertencias del usuario puestas por algún Moderador / Administrador",
                "operationId": "f166a29816206d1717506e5c008dac69",
                "responses": {
                    "202": {
                        "description": "\n    *           $adv1 (string)\n    *           $adv2 (string)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/delete-account": {
            "delete": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Borrar cuenta del usuario",
                "operationId": "336fed4b9d27ea9186a9087a4de22efc",
                "responses": {
                    "202": {
                        "description": "\n    *           Se ha eliminado la cuenta correctamente"
                    },
                    "406": {
                        "description": "\n    *           No se ha podido eliminar la cuenta"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/get-role": {
            "get": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Obtener el rol del usuario",
                "operationId": "5229d25a47a2bd05458e16fe9ff2266b",
                "responses": {
                    "202": {
                        "description": "\n    *           $rol_id (integer)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/edit-user": {
            "put": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Actualizar nombre del usuario",
                "operationId": "863e508776b73cf3968d4e3513aacca0",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "query",
                        "description": "Nombre a actualizar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "apellidos",
                        "in": "query",
                        "description": "Nombre a actualizar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "\n    *           Se ha actualizado correctamente el nombre de usuario"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/edit-email": {
            "put": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Actualizar correo electrónico del usuario",
                "operationId": "f35605ee9e134dd3b5f63b248ad56595",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email a actualizar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Se ha actualizado correctamente el email del usuario"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "http",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}